Use normal cgminer api.
